Migrate rendering logic to internal sympy_input package
